[feat] Member Join 구현& Member 레포지토리 추상화

+++ assignment_3rd/src/main/java/com/example/assignment_3rd/member/entity/Member.java
@@ -8,12 +8,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Member {
-    private int id;
+    private Long id;
     private String name;
     private String password;
 
     @Builder
-    public Member(int id, String name, String password) {
+    public Member(Long id, String name, String password) {
         this.id = id;
         this.name = name;
         this.password = password;
